typed_id provides a typesafe way to have a type <-> id association
There are 3 types introduced. `typed_id` is the base which holds a
single id, this provides a strong relation between type and id.

There are 2 supporting data structures for specific
purposes.
`unordered_typed_ids` holds 1 or more `typed_id` and can be
used to pass multiple ids to a function, preserving the strong type to
id relation.
`ordered_typed_ids` provides a guarantee the contained types are
ordered. It also comes with many convenience methods which can take
advantage of the ordered property.